move new dep files to own dir
