+++ Dockerfile.db
@@ -1,6 +1,6 @@
-FROM postgres:11-alpine
+FROM postgres:13-alpine
 
 RUN mkdir -p /tmp/psql_data/
 
-COPY database/ipr_new_deployment.postgres.dump /tmp/psql_data/ipr_new_deployment.postgres.dump
-COPY database/restore_iprdb_dump.sh /docker-entrypoint-initdb.d/
+COPY database_for_new_deployment/ipr_new_deployment.postgres.dump /tmp/psql_data/ipr_new_deployment.postgres.dump
+COPY database_for_new_deployment/restore_iprdb_dump.sh /docker-entrypoint-initdb.d/
